refactor(matrix): Split MMA verifier into backend interface
- Move existing CPU MMA verification logic into CpuMmaBackend.
- Add IMmaBackend abstraction and wire MmaVerifier through backend->verify().
- Add CUDA backend stub, gateway config flag, and CUDA toolchain detection.
